Create helmholtz2D_wifi
In this example we use the 2D Helmholtz equation to model the propagation of the wifi signal in a room. The room will be a square of 40x40 that consists of 4 exterior walls (which delimitates the square) and 2 interior walls. Also, we consider the hotspot is a disk of a certain radius.