Added a faster algorithm to compute the vandermode matrix and made the divergence based off of inverting the vandermode matrix again. This ruins the accuracy of the gradient
Hello Contributors, I added a file to calculate the vandermode inverse quickly based off of this:https://www.mathworks.com/matlabcentral/fileexchange/127624-fastest-analytic-form-inverse-of-square-vandermonde-matrices ,edited the divergence to be based off of vandermode inversion and test 7 to be like the test I used to generate the figures for the poster presentation. Test 7 fails(it starts off good but the error starts exploding for higher values of k unlike in the poster presentation where it stays mostly stable). This proves that using a fast algorithm for inverting the vandermode matrix kills the accuracy of the calculations for higher orders of k.