feat(fortran): add stub gradient operator & test
This commit adds
* A grad_op_{m,s} module and submodule in src/fortran
* A corresponding test suite in a new "test" subdirectory

The grad_op module & submodule define a grad_op_t derived type
and a stub constructor that can be invoked as grad_op_t(k,m,dx),
where k, m, and dx are the order of accuracy, number of cells,
and step size.  The test simply verifies that the constructor
executes.