fix(divergence,laplacian): def separate types
This commit disambiguates two quantities that behave mathematically
like scalars in some circumstances but are treated differently in
the discrete calculus of mimetic discretizations:

* divergence_1D_t: stored at cell centers & support no operators
* laplacian_1D_t: which extends divergence_1D_t because

In both cases, when the operand is a mathematical scalar, the
results is a scalar.  However, in mimetic discretizations, the
results differ from a scalar quantity in that there are no boundary
values with the agove two types.

With these changes, the following operators map from the listed
operand type to the listed  result type:

* .div.: scalar_1D_t operand, divergence_1D_t result and
* .laplacian.: scalar_1D_t, laplacian_1D_t result,

where for an object `s` of type scalar_1D_t,

  .laplacian. s = .div. (.grad. s)